demo-app/pyproject.toml
[project]
name = "streamlit-app"
version = "1.0.0"
description = ""
authors = [
  {name = "Jonny Daenen", email = "jonny.daenen@dataminded.com"},
  {name = "Pieter Huycke", email = "pieter.huycke@dataminded.com"}
]
maintainers = [
  {name = "Jonny Daenen", email = "jonny.daenen@dataminded.com"},
  {name = "Pieter Huycke", email = "pieter.huycke@dataminded.com"}
]
requires-python = ">=3.12.0, <3.13"
readme = "README.md"
dependencies = [
    "streamlit>=1.40.1,<2",
    "google-cloud-bigquery>=3.27.0,<4",
    "google-generativeai>=0.8.3,<0.9",
    "db-dtypes>=1.3.1,<2",
    "plotly>=6.3.1",
]

[dependency-groups]
dev = [
    "pytest>=8.3.3,<9",
    "pytest-cov>=6.0.0,<7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
